Very simple reconnection strategy
This is the simplest reconnection strategy for the client mode i can come up with.
If the connection is lost the board will reboot since it’s
unclear how the various sockets we use will react to that
event this seems like a safer bet and any solution dealing
with setting up and tearing down sockets in a good way
seems like way more work than it’s worth.

Perhaps this strategy will prove terrible in the long run
but i think this will do as a start.